+++ app/[lang]/showcase/page.tsx
@@ -1,10 +1,8 @@
 import Showcase from "@/components/showcase";
-import { SUPPORTED_LANGUAGES } from "@/i18n/supported-languages";
 
 export const runtime = "edge";
-export const generateStaticParams = () => {
-  return SUPPORTED_LANGUAGES.map((lang) => ({ lang }));
-};
+
+export { generateStaticParams } from "@/i18n";
 
 export default function ShowcasePage() {
   return <Showcase />;
