+++ i18n/generate-static-params.ts
@@ -1,0 +1,5 @@
+import { SUPPORTED_LANGUAGES } from "./supported-languages";
+
+export const generateStaticParams = () => {
+  return SUPPORTED_LANGUAGES.map((lang) => ({ lang }));
+};
