+++ i18n/get-server-translations.ts
@@ -1,6 +1,6 @@
 import { Messages } from "./messages/types";
-import { loadTranslations } from "./load-translations";
-import { interpolate } from "./interpolate";
+import { loadTranslations } from "./internal/load-translations";
+import { interpolate } from "./internal/interpolate";
 import { staticGenerationAsyncStorage } from "next/dist/client/components/static-generation-async-storage.external";
 import { SUPPORTED_LANGUAGES, DEFAULT_LANGUAGE } from "./supported-languages";
 import { getServerCurrentLanguage } from "./get-server-current-language";
