+++ i18n/internal/load-translations.ts
@@ -1,5 +1,5 @@
-import { Messages } from "./messages/types";
-import { SUPPORTED_LANGUAGES, SupportedLanguage } from "./supported-languages";
+import type { Messages } from "../messages/types";
+import { SUPPORTED_LANGUAGES, SupportedLanguage } from "../supported-languages";
 
 export async function loadTranslations(lang: string): Promise<Messages> {
   let messages: Messages;
@@ -9,7 +9,7 @@ export async function loadTranslations(lang: string): Promise<Messages> {
   const languageToLoad = isSupported ? lang : "en"; // 지원되지 않는 언어의 경우 기본 언어로 폴백
 
   try {
-    messages = await import(`./messages/${languageToLoad}`).then(
+    messages = await import(`../messages/${languageToLoad}`).then(
       (module) => module.default,
     );
   } catch (error) {
@@ -19,7 +19,9 @@ export async function loadTranslations(lang: string): Promise<Messages> {
     );
     // 메시지 로드 실패 시 기본 언어로 폴백
     if (languageToLoad !== "en") {
-      messages = await import("./messages/en").then((module) => module.default);
+      messages = await import("../messages/en").then(
+        (module) => module.default,
+      );
     } else {
       throw new Error("Failed to load default language messages.");
     }
